Revert "Moved entities to app.core.models"
This reverts commit b92d235aeba6f20ee599a2bca1030c06873a0b9b.
