+++ application/backend/app/core/models/__init__.py
@@ -12,9 +12,6 @@ from .base import (
     RequiresID,
     RequiresName,
 )
-from .dataset_item import DatasetItem, DatasetItemAnnotation, DatasetItemFormat, DatasetItemSubset
-from .label import LabelReference
-from .shape import FullImage, Point, Polygon, Rectangle, Shape
 from .task_type import TaskType
 
 __all__ = [
@@ -22,20 +19,10 @@ __all__ = [
     "BaseModel",
     "BaseRequiredIDModel",
     "BaseRequiredIDNameModel",
-    "DatasetItem",
-    "DatasetItemAnnotation",
-    "DatasetItemFormat",
-    "DatasetItemSubset",
-    "FullImage",
     "HasID",
     "HasName",
-    "LabelReference",
     "Pagination",
-    "Point",
-    "Polygon",
-    "Rectangle",
     "RequiresID",
     "RequiresName",
-    "Shape",
     "TaskType",
 ]
