+++ application/backend/app/entities/base.py
@@ -1,0 +1,7 @@
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+from pydantic import BaseModel, ConfigDict
+
+
+class BaseEntity(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
