+++ application/backend/app/services/dataset_service.py
@@ -16,16 +16,9 @@ import numpy as np
 from PIL import Image, UnidentifiedImageError
 from sqlalchemy.orm import Session
 
-from app.core.models import (
-    DatasetItem,
-    DatasetItemAnnotation,
-    DatasetItemSubset,
-    FullImage,
-    Polygon,
-    Rectangle,
-    TaskType,
-)
+from app.core.models import TaskType
 from app.db.schema import DatasetItemDB
+from app.entities import DatasetItem, DatasetItemAnnotation, DatasetItemSubset, FullImage, Polygon, Rectangle
 from app.repositories import DatasetItemRepository
 from app.schemas.label import LabelBase
 from app.schemas.project import ProjectBase, ProjectView
